+++ ui/src/components/Publication/Visualization/index.tsx
