+++ ui/src/components/index.tsx
@@ -65,7 +65,7 @@ export { default as PublicationSidebarCardSections } from './Publication/Sidebar
 export { default as PublicationSimpleResult } from './Publication/SimpleResult';
 export { default as PublicationSkeleton } from './Publication/Skeleton';
 export { default as PublicationVisualChain } from './Publication/VisualChain';
-export { default as PublicationVisulization } from './Publication/Visulization';
+export { default as PublicationVisulization } from './Publication/Visualization';
 export { default as RenderLinks } from './RenderLinks';
 export { default as RequiredIndicator } from './RequiredIndicator';
 export { default as RORForm } from './RORForm';
